# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Medclara Backend API
  description: |
    # Medclara Clinical Documentation Platform API
    
    Complete API specification for the Medclara backend, including:
    - **Authentication** - User login and session management
    - **Patients** - Patient data management
    - **Templates** - Clinical note templates
    - **Sessions** - Scribe session management
    - **Recordings** - Audio recording upload and processing
    - **Chunked Upload** - Large file upload with resumable chunks
    - **Server-Sent Events (SSE)** - Real-time processing status updates
    - **Notes** - Clinical note creation and management
    - **Workspace** - Scribe workspace API
    
    ## Real-Time Processing with SSE
    
    The Medclara API uses Server-Sent Events (SSE) for real-time recording processing updates.
    This provides 99% fewer network requests and 20x faster response times compared to polling.
    
    ### Complete Recording Workflow
    1. **POST** `/api/recordings/chunked/init` - Initialize upload session
    2. **POST** `/api/recordings/chunked/upload` - Upload audio chunks (repeat)
    3. **POST** `/api/recordings/chunked/finalize` - Start processing
    4. **GET** `/api/recordings/{recordingId}/stream` - Stream processing events (SSE)
    5. Process events: `processing` → `completed` or `failed`
    
    ### Benefits of SSE Over Polling
    - **99% fewer requests** - One persistent connection vs. polling every 500ms
...
    
  version: '1.0.0'
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code for programmatic handling
          examples:
            - unauthorized
            - not_found
            - invalid_request
            - server_error
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...
    Note:
      type: object
      required:
        - id
        - patient_id
        - title
        - content
        - status
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        recording_id:
          type: string
          format: uuid
        encounter_id:
          type: string
        title:
          type: string
        content:
          type: string
          description: Note content
        note_type:
          type: string
          enum: [scribe, clinical, followup, general]
        status:
          type: string
          enum: [draft, completed, signed, archived]
        scribe_page_id:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        extracted_data:
          type: object
          description: AI-extracted structured data
          additionalProperties: true
        extraction_method:
          type: string
          enum: [manual, ai, hybrid]
        tags:
          type: array
          items:
            type: string
        is_signed:
          type: boolean
        signed_at:
          type: string
          format: date-time
        signed_by:
          type: string
          format: uuid
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
    CreateNoteRequest:
      type: object
      required:
        - patient_id
        - title
        - content
      properties:
        patient_id:
          type: string
          format: uuid
        recording_id:
          type: string
          format: uuid
        encounter_id:
          type: string
        title:
          type: string
        content:
          type: string
        note_type:
          type: string
          enum: [scribe, clinical, followup, general]
        scribe_page_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        extracted_data:
          type: object
          additionalProperties: true
        extraction_method:
          type: string
          enum: [manual, ai, hybrid]
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        note_type:
          type: string
          enum: [scribe, clinical, followup, general]
        status:
          type: string
          enum: [draft, completed, signed, archived]
        template_id:
          type: string
          format: uuid
        extracted_data:
          type: object
          additionalProperties: true
        extraction_method:
          type: string
          enum: [manual, ai, hybrid]
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true

    # ========================================================================
    # Workspace Schemas
    # ========================================================================
...
paths:
...
  /api/notes:
    post:
      tags:
        - Notes
      summary: Create note
      description: Create a new clinical note
      operationId: createNote
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Notes
      summary: List notes by patient
      description: Retrieve notes for a specific patient
      operationId: listNotesByPatient
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Note list
          content:
            application/json:
              schema:
                type: object
                properties:
                  notes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
                  total:
                    type: integer
...
  /api/notes/{noteId}:
    get:
      tags:
        - Notes
      summary: Get note
      description: Retrieve specific note
      operationId: getNote
      security:
        - BearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Note details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Notes
      summary: Update note
      description: Update a note
      operationId: updateNote
      security:
        - BearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '200':
          description: Note updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Notes
      summary: Delete note
      description: Delete a note
      operationId: deleteNote
      security:
        - BearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Note deleted
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notes/{noteId}/sign:
    post:
      tags:
        - Notes
      summary: Sign note
      description: Sign a note (mark as officially completed)
      operationId: signNote
      security:
        - BearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Note signed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  signedAt:
                    type: string
                    format: date-time
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================================
  # WORKSPACE ENDPOINTS
  # ============================================================================
...
security:
  - BearerAuth: []